Implement adding card
