+++ lib/util/diary_entry.dart
@@ -1,0 +1,11 @@
+class DiaryEntry {
+  final String title;
+  final String username;
+  final String description;
+
+  DiaryEntry({
+    required this.title,
+    required this.username,
+    required this.description,
+  });
+}
